+++ agent/listener/database.py
@@ -175,6 +175,16 @@ class DB(object):
             for col in columns:
                 check[col[0]] = obj[i]
                 i += 1
+
+            # Process output types
+            output = check['output'].split("\n")
+            check['output'] = output[0]
+
+            # Give long output if it exists
+            check['longoutput'] = ''
+            if len(output) > 1:
+                check['longoutput'] = "<br>".join(output[1:])
+
             checks.append(check)
 
         return checks